requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==10.0.1
pytesseract==0.3.10
PyPDF2==3.0.1
python-dateutil==2.8.2
bson==0.5.10
